refactor stage
