+++ pkg/common/stage/stage_test.go
@@ -1,0 +1,37 @@
+// Copyright 2024 Matrix Origin
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package stage
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestS3ServiceProvider(t *testing.T) {
+	protocol, err := getS3ServiceFromProvider("cos")
+	require.Nil(t, err)
+	assert.Equal(t, protocol, "s3")
+
+	protocol, err = getS3ServiceFromProvider("amazon")
+	require.Nil(t, err)
+	assert.Equal(t, protocol, "s3")
+
+	protocol, err = getS3ServiceFromProvider("minio")
+	require.Nil(t, err)
+	assert.Equal(t, protocol, "minio")
+
+}
